requirements.txt
#### Bot lib ####
python-telegram-bot>=13.0.0
telethon
#################
aiohttp>=2.2.5
bleach
emoji
future
googletrans>=3.1.0a0
hurry.filesize
markdown2
Pillow>=4.2.0
psycopg2-binary
pydantic
python-rapidjson
python-dotenv
PyYAML
requests>=2.20.0
sqlalchemy
tldextract
wikipedia
covid
